cli: split per sub-command
